+++ CKGL/src/Math/Math.cs
@@ -57,7 +57,7 @@
 			//return i;
 		}
 
-		public static float Clamp(float f, int min, int max)
+		public static float Clamp(float f, float min, float max)
 		{
 			return (f > max ? max : (f < min ? min : f));
